v9.0 refactor UserDbStorage + UserService

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 import ru.yandex.practicum.filmorate.validator.UserValidator;
@@ -22,6 +23,9 @@ public class UserService {
     }
 
     public User addUser(User user) {
+        if (userStorage.containsUser(user)) {
+            throw new NotFoundException("This user already exists");
+        }
         if (user.getName() == null || user.getName().isBlank()) {
             user.setName(user.getLogin());
         }
@@ -47,22 +51,37 @@ public class UserService {
     }
 
     public void addFriend(Integer userId, Integer friendId) {
+        if (!userStorage.containsUser(userId) || !userStorage.containsUser(friendId)) {
+            throw new NotFoundException("Users not found.");
+        }
         userStorage.addFriend(userId, friendId);
     }
 
     public void deleteFriend(Integer userId, Integer friendId) {
+        if (!userStorage.containsUser(userId) || !userStorage.containsUser(friendId)) {
+            throw new NotFoundException("you cannot delete a non-existent user");
+        }
         userStorage.deleteFriend(userId, friendId);
     }
 
     public List<User> getCommonFriends(Integer userId, Integer friendId) {
+        if (!userStorage.containsUser(userId) || !userStorage.containsUser(friendId)) {
+            throw new NotFoundException("Unable to get friends list of non-existent user");
+        }
         return  userStorage.getCommonFriends(userId, friendId);
     }
 
     public List<User> getFriends(Integer friendId) {
+        if (!userStorage.containsUser(friendId)) {
+            throw new NotFoundException("Unable to get friends list of non-existent user");
+        }
         return userStorage.getFriends(friendId);
     }
 
     public User getUser(Integer userId) {
+        if (!userStorage.containsUser(userId)) {
+            throw new NotFoundException("No such user exists");
+        }
         return userStorage.getUser(userId);
     }
 }
